src/requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
filelock==3.9.0
Flask==2.2.3
Flask-Login==0.6.2
identify==2.5.18
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
nodeenv==1.7.0
platformdirs==3.0.0
pre-commit==3.1.1
pymongo==4.3.3
pysolr==3.9.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
url-normalize==1.4.3
urllib3==1.26.14
virtualenv==20.20.0
waitress==2.1.2
Werkzeug==2.2.3
